+++ src/ui/service/crossterm/mod.rs
@@ -1,0 +1,20 @@
+// devela::ui::service::crossterm
+//
+//! `crossterm` Ui service.
+//
+
+mod service;
+// mod events;
+
+crate::items! { // structural access: _mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods {
+        pub use super::service::*;
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}